add delay after done punish and start drawing on new turn
