+++ ui/main.go
@@ -172,12 +172,12 @@ func (g *Game) ChangeState(stateName string) {
 		// kk := cards.NewRookieMage(mm.defaultGamestate)
 		// slimeRoom := cards.NewSlimeRoom(mm.defaultGamestate)
 		// boulder := cards.NewBoulderTrap(mm.defaultGamestate)
-		// spikeFloor := cards.NewDamageEndturnCurse(mm.defaultGamestate) //cards.NewSpikeFloor(mm.defaultGamestate)
+		spikeFloor := cards.NewTorchtail(mm.defaultGamestate) //cards.NewDamageEndturnCurse(mm.defaultGamestate) //cards.NewSpikeFloor(mm.defaultGamestate)
 		// lair := cards.NewGoblinSmallLairArea(mm.defaultGamestate)
 		// heal := item.NewHealingPotion(defaultGamestate)
-		// ll := append(mm.defaultGamestate.CardsInCenterDeck.List()[:3], &spikeFloor)
-		// rest := mm.defaultGamestate.CardsInCenterDeck.List()[4:]
-		// mm.defaultGamestate.CardsInCenterDeck.SetList(append(ll, rest...))
+		ll := append([]cards.Card{}, &spikeFloor)
+		rest := mm.defaultGamestate.CardsInCenterDeck.List()
+		mm.defaultGamestate.CardsInCenterDeck.SetList(append(ll, rest...))
 		// mm.defaultGamestate.ItemCards = append(mm.defaultGamestate.ItemCards, &heal)
 		// mm.defaultGamestate.CardsInHand = append(mm.defaultGamestate.CardsInHand, &spikeFloor)
 		// mm.defaultGamestate.CardsInDeck.Stack(&wl)
